:truck: Move field definition getter hook into file and stop exporting get field def
